syntax fix type int

+++ faceTracking.py
@@ -100,8 +100,8 @@ def faceTracking(rawVideo):
             #get the bounding box for the current face
             bboxOfCurrFace = newbbox[facei, :, :]
             #get the positions of the two corners for the bounding box of the current face
-            first = bboxOfCurrFace[0,:]
-            second = bboxOfCurrFace[3,:]
+            first = bboxOfCurrFace[0,:].astype(int)
+            second = bboxOfCurrFace[3,:].astype(int)
             img2WithBoundingBox = cv2.rectangle(img2WithBoundingBox, (first[0],first[1]), (second[0],second[1]), (255,0,0))
 
         #add img2 to the output matrix
